requirements.txt
Flask==2.1.1
pandas==1.3.3
requests==2.26.0
plotly==5.3.1
d3==6.2.0
urllib3==1.26.6
matplotlib==3.4.3
seaborn==0.11.2
facebook-sdk==3.0.0
shopify==14.1.0
shopifyapi==8.0.0